find companion by id

+++ src/main/java/com/bayer/data/dao/AddressDAO.java
@@ -53,7 +53,7 @@ public class AddressDAO implements CRUD<Address>{
     }
 
     @Override
-    public Address findById(int id) {
+    public Address findById(long id) {
         return null;
     }
